Add files via upload
A simple Library Management System implemented in C that allows users to manage a collection of books. This program features functionalities to add books, view all books, and search for a book by title. It serves as a foundational project for understanding basic data management and console input/output in C.

Features:

Add new books with title and author details.
View a list of all books in the library.
Search for books by title.